can read accurate stats using threads

+++ src/com/ikhokha/techcheck/CommentReport.java
@@ -1,9 +1,9 @@
 package com.ikhokha.techcheck;
 
 public class CommentReport {
-    public int Movers;
-    public int Shakers;
-    public int Shorter;
-    public int Questions;
-    public int Spam;
+    public long Movers;
+    public long Shakers;
+    public long Shorter;
+    public long Questions;
+    public long Spam;
 }
